+++ player.go
@@ -1,7 +1,7 @@
 package main
 
 type player struct {
-	id           int
+	id           int64
 	currentBlock int32
 	score        int
 	active       bool
@@ -22,7 +22,7 @@ func newColor(r, g, b uint8) (rgb rgb_color) {
 	return rgb
 }
 
-func newPlayer(id int, rgb rgb_color) (p player) {
+func newPlayer(id int64, rgb rgb_color) (p player) {
 	p.id = 1
 	p.currentBlock = -1
 	p.score = 0
